+++ client/src/io/worxbend/gitea4s/api/IssuesApi.scala
@@ -2,7 +2,7 @@ package io.worxbend.gitea4s.api
 
 import io.worxbend.gitea4s.error.GiteaError
 import io.worxbend.gitea4s.http.IssueListParams
-import io.worxbend.gitea4s.model.Issue
+import io.worxbend.gitea4s.model.{CreateIssue, Issue}
 import zio.IO
 import zio.stream.ZStream
 
@@ -14,3 +14,5 @@ trait IssuesApi:
       repo: String,
       params: IssueListParams = IssueListParams.default
   ): ZStream[Any, GiteaError, Issue]
+
+  def create(owner: String, repo: String, body: CreateIssue): IO[GiteaError, Issue]